clean up and add proper docs.
